pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vision-embeddings"
version = "0.1.0"
description = "Extract, cache, and upload vision encoder embeddings from VLM training datasets"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "Gokay Aydogan" }]
keywords = ["vision", "embeddings", "vlm", "siglip", "clip", "vjepa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.1.0",
    "transformers>=4.48.0",
    "datasets>=3.0.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.20.0",
    "Pillow>=10.0.0",
    "tqdm>=4.65.0",
    "accelerate>=0.25.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
vjepa = ["timm", "einops"]
all = ["timm", "einops"]

[project.scripts]
vision-embeddings = "vision_embeddings.__main__:main"

[project.urls]
Repository = "https://github.com/gokayfem/vision-embeddings"
